imgtool: Improve key type checks and invalid pass-phrase handling
Key type checking moved to separate function and added support for
ed25519, enckey is enforced to be a public key

Invalid pass-phrase was not reachable in methods as in that case an
exception raised instead of returning "None".
load_key function improved by adding handling for invalid pass-phrase
and FileNotFound exception.

Signed-off-by: Rustam Ismayilov <rustam.ismayilov@arm.com>
Change-Id: I0caa0a6100fc95c8339403e991d6de0337c7a725