boot: zephyr: don't disable `MULTITHREADING`
The documentation on the `MULTITHREADING` symbol states that
```
If you know for sure that your hardware will work, you can default it to
n here.
```
Defaulting the entire Nordic user base to `n` here is a huge statement
that all Nordic hardware will work without `MULTITHREADING`, which is
somewhat undermined by all of the board overlays that have to turn it
back on again in order to get basic features working (USB recovery,
images on SPI or QSPI flash).

Disable it again in the `nrf52_minimal_footprint.conf` sample, as that
is obviously intended to showcase the minimal footprint possible.

Leave it disabled for the `SOC_SERIES_NRF51X`, as the default boot
sector partitions for that SoC are small enough to require it.

Signed-off-by: Jordan Yates <jordan@embeint.com>