boot: zephyr: remove explicit `NORDIC_QSPI_NOR` control
Remove explicit control over `NORDIC_QSPI_NOR` from board overlays
(including some boards that didn't have QPI flash anyway). The new
behaviour is that if the QSPI flash node exists in devicetree, it will
be compiled into the bootloader.

Options for disabling this include:
   * Setting the devicetree node to `status="disabled"` for MCUboot
   * Adding a default in the board definition
```
configdefault NORDIC_QSPI_NOR
  default n if MCUBOOT
```

Explicitly disabling the symbol in the MCUboot repository makes it very
painful for users of the board to setup an alternate flash partition
scheme (secondary on QSPI flash), as it requires maintaining a fork of
MCUboot or board specific overrides at compile-time.

Signed-off-by: Jordan Yates <jordan@embeint.com>