boot: zephyr: Allow disabling NRFX_WDT on nRF devices
Allow disabling feeding the watchdog on nrf devices directly
using the nrfx_wdt functions.
The imply on NRFX_WDT cannot be disabled since NRFX_WDT is a
promptless symbol and cannot be assigned in a config fragment.
Not even deleting the wdt0 node in a devicetree overlay works
since the configuration takes predence and ends in a #error without
any nrfx_wdt instances.

To fix this create new Kconfig option with prompt and change the imply
to this option. That way the option can be assigned to off by the user.

This is needed to support the use case of an external watchdog
controlled using the zephyr watchdog API instead of the wdt peripheral.

Signed-off-by: Joakim Andersson <joerchan@gmail.com>