zephyr: Fixing Kconfig dependency for SHA512
The SHA512_ALLOW Kconfig has been added to allow signature
algorithms to select which SHA they support. Unfortunately
it has been given dependency on PSA crypto, which now is problematic
because if signature algorithm wants to indicate that it allows
SHA512 it immediately becomes dependent on PSA crypto.

The commit removes the dependency.

Signed-off-by: Dominik Ermel <dominik.ermel@nordicsemi.no>