boot/bootutil: Get bootloader state
Add `boot_get_loader_state()` to allow one have a reference to current
bootloader state. While this state is internal - and struct is opaque -
it will be important to be able to "pass it around" when using hooks
that will be introduced in future patches.

While at it, make Zephyr single loader also expose it.

Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>