boot: New boot_go hook
This new hook allows external modules to override the choice of which
image to boot - it can be used, for instance, to load an image based on
some GPIO state.

If the hook returns FIH_BOOT_HOOK_REGULAR, then normal platform
boot_go() will be run to choose image to boot.

This hook is shielded by a different configuration than other hooks -
MCUBOOT_BOOT_GO_HOOKS, so that it can be enabled independently from
other available hooks.

Support for this new hook was added to Zephyr port.

Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>