samples/runtime-source: Make it buildable on twister
This sample has two parts: hooks for an mcuboot instance and a target
application. For the first, a new entry at boot/zephyr/sample.yaml is
added as well as a corresponding change at boot/zephyr/CMakeLists.txt to
include the hook as a module. For the second, added a sample.yaml as
well as invoke it from the github workflow.

Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>