boot: bootutil: Swap offset: Do not fail on header in wrong sector
Swap offset does not expect header in first sector/page of secondary
slot, rather in secondary, and only attempts to swap an image that
starts at second sector.
This commit demotes condition when header is found in first page
to just a warning and allows boot to proceed, as long as there
is a header is second slot.

Signed-off-by: Dominik Ermel <dominik.ermel@nordicsemi.no>