boot: bootutil: Ensure the whole trailer is erased when it is large
When using swap-scratch and the trailer is too large to fit in a single
sector, might not be fully erased in both the primary and the secondary
slot if the last sector containing firmware data also contains part of
the trailer. Indeed, in that case, it might happen that only the first
part of the trailer was erased, causing issues e.g. when attempting to
rewrite the trailer in the primary slot during the upgrade.

Signed-off-by: Thomas Altenbach <thomas.altenbach@legrand.com>