sim: Fix incorrect trailer size computation for overwrite-only
For the overwrite-only upgrade strategy, the trailer size computed by
the simulator and used to determine the maximum image size was not
correct. This commit fixes the issue.

Having an underestimated trailer size was causing the
'oversized_secondary_slot' to fail since the previous commit, because
the oversized images are now generated to have the smallest possible
size.

Signed-off-by: Thomas Altenbach <thomas.altenbach@legrand.com>