boot/mynewt: remove redundant mbedtls flag and enable CTR
Now flag defining config file for MbedTLS is global
(see https://github.com/apache/mynewt-core/pull/3394),
so we do not have to include the same flag in bootutil package.
This also enables MBEDTLS_CIPHER_MODE_CTR in boot_serial test
package, as it is used in unit tests and it is disabled by default.

Signed-off-by: Michal Gorecki <michal.gorecki@codecoup.pl>