bootutil: Make boot_get_first_trailer_sector available for swap-move
This routine will now also be useful for the swap-move strategy. The
trailer size is removed from the list of parameters as it can be easily
obtained using the 'state' parameter and it will be more convenient, for
the swap-move strategy, not to have to compute it manually before
calling 'boot_get_first_trailer_sector'.

Signed-off-by: Thomas Altenbach <thomas.altenbach@legrand.com>