bootutil: swap-move: Update maximum size to allow unaligned trailers
This commit updates the computation of the maximum firmware image size
for the swap-move strategy to allow the first sector containing part of
the trailer to also hold firmware data. This means it is no more
necessary to allocate a sector-aligned area for the trailer when using
swap-move, the area to allocate is now simply equal to the maximum
trailer size.

Signed-off-by: Thomas Altenbach <thomas.altenbach@legrand.com>