bootutil: swap-move: Avoid rewriting the secondary trailer on revert
When using the swap-move strategy, at the very beginning of the revert
process, the secondary trailer was rewritten to make the revert look
like a permanent upgrade in case an unfortunate reset occurs when
rewriting the primary trailer. This was possible because the assumption
was that no sector contained both part of the firmware and part of the
trailer.

To relax this assumption, it is necessary to avoid having to rewrite the
secondary trailer at the start of the revert process, since that could
also erase firmware data. The solution chosen is to write, during the
upgrade process, a "fallback trailer" at the end of the last sector that
is not containing part of the slot's trailer. The fallback trailer only
contains the magic field. Depending on the size of the trailer, the
maximum firmware image size has to be reduced a bit (by at most 8 bytes)
to ensure the fallback trailer can be written. During a revert, if a
reboot occurs while the primary trailer is rewritten, the presence of
the fallback trailer will make possible to determine that a revert was
in progress and has to be resumed.

Signed-off-by: Thomas Altenbach <thomas.altenbach@legrand.com>