sim: Fix largest image computation for swap-move and swap-offset
For swap-move and swap-offset strategies, the size of the images used
for running the tests was hardcoded and was smaller than the maximum
possible size. This was done this way because the logic used to compute
the maximum image size was not taking into account the padding that is
needed when using those strategies. This commit fixes the issue, making
now possible to run the tests with images having the maximum possible
size.

Signed-off-by: Thomas Altenbach <thomas.altenbach@legrand.com>