sim: swap-move: Allow firmware data in the first trailer sector
When using the swap-move strategy, the simulator now doesn't compute
anymore the maximum image size so the area allocated to the trailer has
a size that is a multiple of the sector size. Indeed, the swap-move
strategy now supports having part of the firmware in the first sector
holding trailer data, allowing larger images.

Signed-off-by: Thomas Altenbach <thomas.altenbach@legrand.com>