swap_move: correct appsize calcs and warning
app_max_sectors() is always off by one in its calculation.
This corrects that calculation as well as the warning which
should only be comparing the two slot sector counts and
checking to see if they are different by one sector for
the swap sector. The size of the application is not relevant
in that warning for optimal sector distribution.

Additionally app_max_size() suffers similarly in that it
is not taking into account the fact that the secondary
slot does also contain a trailer.

Finally makes a minor addition to the design document which
further clarifies the primary and secondary partition size
differences.

Signed-off-by: Robert Wolff <bob.wolff68@gmail.com>