boot: zephyr: move CDC ACM serial to the new device stack
The new stack comes with a higher demand for flash. We need to adapt the
overlay file for the platform used for testing. The legacy stack is in
the process of being deprecated. The NRF52840 dongle should not enable
the legacy stack by default. The Zephyr's new stack provides helper code
to use CDC ACM as the serial backend. The only required change to the
serial adapter code is to get the USB device instance and enable it.

Signed-off-by: Johann Fischer <johann.fischer@nordicsemi.no>