boot: Remove now superfluous start_off parameters
When swap-offset is used, multiple routines called directly or
indirectly by bs_list or bs_set needed to know the start offset of the
image in the secondary slot. When a valid bootloader state is available,
this value is simply retrieved from the state. Since bs_list and bs_set
had no valid state, a start_off parameter was added to all those
routines to obtain the start offset without using the bootloader state.

This is no longer needed, and since the start_off parameters were used
only by bs_list and bs_set, they can be removed, which make the code
simpler and easier to read in numerous places.

Signed-off-by: Thomas Altenbach <thomas.altenbach@legrand.com>