boot: zephyr: adjust sample.yaml to pass USB deprecation process
The legacy USB device stack is in the process of being deprecated. In
order to pass CI in the Zephyr Project, we need to adjust the test case
settings in sample.yaml. Clean up both test cases and add a Kconfig
option that allows them to pass Twitter runs despite deprecation
warnings and flash overflows.

Signed-off-by: Johann Fischer <johann.fischer@nordicsemi.no>