scripts: imgtool: Add script to combine multiple image as single
MCUBoot multi image support mode requires each image
be procedded individually that requires multiple signature
check during boot. If there be 4 imageis it requires 4 times
signature validatation. This feature increase boot time.
Depend on the project requirement long boot time may not be accepted.

To provide a solution for this case we propose to
- Generate each image as regular mcuboot format
- Concatanate them and re generate a main mcuboot image. As below

-------------------------
|         Header        |
-------------------------
|        Image-1        |
|  (Header+Data+Footer) |
-------------------------
|        Image-2        |
|  (Header+Data+Footer) |
-------------------------
|        .....          |
-------------------------
|         Footer        |
-------------------------

During boot time if top level image be validated sub image can be
just copied to the target location without to recheck them.

There will be two commit to implement this feature
1- Add a script that combine images
2- Update mcuboot source code to process subimages

This commit is for #1, it adds a script which called as
combine_images.py. Any imgtool parameters can be passed to the script
It will directly pass them to the imgtool.

Usage:
python combine_images.py --config combine_images.yaml --imgtool imgtool --output <outfolder>

combine_images.yaml file is added as reference file it need to be
updated as per of the project need.

Signed-off-by: Sadik Ozer <sadik.ozer@analog.com>
Signed-off-by: Michael Eskowitz <Michael.Eskowitz@analog.com>