scripts: imgtool: Add script to combine multiple image as single
MCUBoot multi image support mode requires each image
be procedded individually that requires multiple signature
check during boot. If there be 4 image it will require 4 times
signature validatation. This feature increase boot time and
depend on the project long boot time may not be expected.

To provide a solution for this case we propose to
- Generate each image as regular
- Concatanate them and re generate a main mcuboot image.

-------------------------
|         Header        |
-------------------------
|  SubImage (optional)  |
|  (Header+Data+Footer) |
-------------------------
|  SubImage (optional)  |
|  (Header+Data+Footer) |
-------------------------
|        .....          |
-------------------------
|         Footer        |
-------------------------

During boot time if top level image be validated sub image can be
just copied to the target location without to recheck them.

To provide this feature there will be two commit
1- Provide a script that combine images
2- Update mcuboot source code to process subimages

This commit is for #1, in this commit we are adding
a script which called as combine_images.py.

The usage of the script is:
python combine_images.py --config combine_images.yaml --imgtool imgtool --output <outfolder>

combine_images.yaml file is added as reference file it need to be
updated as per of the project need.

Signed-off-by: Sadik Ozer <sadik.ozer@analog.com>
Signed-off-by: Michael Eskowitz <Michael.Eskowitz@analog.com>