loader: Enforce version equality in slot deps
In case of slotted dependency, interpret the version as a strict
requirement, not just as a minimum compatible version.

This change fixes a problem of incorrect dependency resolution for
fully dependent slots (a,s0 -> b,s0 and b,s0 -> a,s0), because the
satisfaction of both of them will be checked in an image with lower
index, which is required for the current algorithm to work correctly.

Signed-off-by: Tomasz Chyrowicz <tomasz.chyrowicz@nordicsemi.no>