gitignore: Ignore some files related to vscode
The clangd plugin in vscode is useful for working on the code. However,
in order to function, especially with numerous feature flags and such,
it needs a `compile-commands.json` file.  This can be generated for a
given build with something like:

    bear --append -- cargo test --features sig-ecdsa

which will create this file.

The clangd plugin also creates a `.cache` directory to hold cached
versions of its analysis.

Support this usage by adding both of these to the .gitignore

Signed-off-by: David Brown <david.brown@linaro.org>