rust: Make workspace more convenient with default
Set the default-members in the top-level `Cargo.toml` file.  This tells
cargo what to do if it is invoked at the top level.

Using this allows builds with the simulate to be done directly from the
top level directory without having to manually cd to the sim directory.

Combined with `bear`, this makes development with vscode much more
handy.

Signed-off-by: David Brown <david.brown@linaro.org>