bootutil: Temporarly drop mem cleanup from boot_state_clear
Temporal removal of memset clear of boot_sector_clear that
was causing boot to stop.
The problem happens due to fill_rsp assigning pointer to header
stored in cleared boot_state object, and the memset makes it inaccessible.

Once change where header is copied to rsp the memset should be brought
back as the boot_loader state should be completed before we pass
further in execution.

Fixes #2535

Signed-off-by: Dominik Ermel <dominik.ermel@nordicsemi.no>